[ADD] fractol: almost done

+++ ll3d/ll3d.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ll3d.h                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sbednar <sbednar@student.42.fr>            +#+  +:+       +#+        */
+/*   By: sbednar <sbednar@student.fr.42>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/19 14:56:09 by edraugr-          #+#    #+#             */
-/*   Updated: 2019/02/21 04:42:57 by sbednar          ###   ########.fr       */
+/*   Updated: 2019/03/04 18:01:26 by sbednar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -75,22 +75,22 @@ typedef struct			s_vec2
 
 typedef struct			s_vec3
 {
-	float				x;
-	float				y;
-	float				z;
+	long double				x;
+	long double				y;
+	long double				z;
 }						t_vec3;
 
 typedef struct			s_vec4
 {
-	float				x;
-	float				y;
-	float				z;
-	float				w;
+	long double				x;
+	long double				y;
+	long double				z;
+	long double				w;
 }						t_vec4;
 
 typedef struct			s_mtx4
 {
-	float				val[4][4];
+	long double				val[4][4];
 }						t_mtx4;
 
 typedef struct			s_image
@@ -118,10 +118,10 @@ typedef struct			s_cam
 	t_mtx4				*cam;
 	t_vec3				*piv;
 	int					mode;
-	float				speed;
-	float				ppu;
-	float				sence;
-	float				focus;
+	long double				speed;
+	long double				ppu;
+	long double				sence;
+	long double				focus;
 	struct s_cam		*add;
 }						t_cam;
 
@@ -146,59 +146,59 @@ t_vec2					*vec2_copy(const t_vec2 *v);
 void					vec2_set(t_vec2 *v, const int x, const int y);
 
 t_vec3					*vec3_init(void);
-t_vec3					*vec3_new(const float x, const float y, const float z);
+t_vec3					*vec3_new(const long double x, const long double y, const long double z);
 t_vec3					*vec3_sum(const t_vec3 *v1, const t_vec3 *v2);
-t_vec3					*vec3_mult(const t_vec3 *v, const float s);
+t_vec3					*vec3_mult(const t_vec3 *v, const long double s);
 t_vec3					*vec3_copy(const t_vec3 *v);
-void					vec3_set(t_vec3 *v, const float x, const float y,
-						const float z);
+void					vec3_set(t_vec3 *v, const long double x, const long double y,
+						const long double z);
 
 t_vec4					*vec4_init(void);
-t_vec4					*vec4_new(const float x, const float y, const float z,
-						const float w);
+t_vec4					*vec4_new(const long double x, const long double y, const long double z,
+						const long double w);
 t_vec4					*vec4_sum(const t_vec4 *v1, const t_vec4 *v2);
-t_vec4					*vec4_mult(const t_vec4 *v, const float s);
+t_vec4					*vec4_mult(const t_vec4 *v, const long double s);
 t_vec4					*vec4_copy(const t_vec4 *v);
 t_vec4					**dots4_cpy(const t_vec4 **src, const int xc,
 						const int yc);
 
 int						ft_abs(const int x);
-int						ft_sign(const float a);
-float					ft_fabs(const float x);
+int						ft_sign(const long double a);
+long double					ft_fabs(const long double x);
 void					ft_swap(int *a, int *b);
-float					ft_fmin(const float a, const float b);
+long double					ft_fmin(const long double a, const long double b);
 int						ft_min(const int a, const int b);
-float					ft_fmax(const float a, const float b);
+long double					ft_fmax(const long double a, const long double b);
 int						ft_max(const int a, const int b);
 
-t_mtx4					*mtx4_init(const float val);
+t_mtx4					*mtx4_init(const long double val);
 void					mtx4_clear(t_mtx4 *mtx);
 t_mtx4					*mtx4_copy(const t_mtx4 *src);
 t_mtx4					*mtx4x3_mult(const t_mtx4 *a, const t_mtx4 *b,
 						t_mtx4 *dst);
 t_mtx4					*mtx4x4_mult(const t_mtx4 *a, const t_mtx4 *b,
 						t_mtx4 *dst);
-t_mtx4					*mtx4_get_rotation(const float x, const float y,
-						const float z, const float a);
+t_mtx4					*mtx4_get_rotation(const long double x, const long double y,
+						const long double z, const long double a);
 
 t_cam					*camera_init(void);
-void					camera_move_x(t_cam *cam, const float a);
-void					camera_move_y(t_cam *cam, const float a);
-void					camera_move_z(t_cam *cam, const float a);
-void					camera_rotate_x(t_cam *cam, const float a);
-void					camera_rotate_y(t_cam *cam, const float a);
+void					camera_move_x(t_cam *cam, const long double a);
+void					camera_move_y(t_cam *cam, const long double a);
+void					camera_move_z(t_cam *cam, const long double a);
+void					camera_rotate_x(t_cam *cam, const long double a);
+void					camera_rotate_y(t_cam *cam, const long double a);
 void					camera_rotate_custom(t_cam *cam, const t_vec3 *p,
-						const t_vec3 *axis, const float a);
+						const t_vec3 *axis, const long double a);
 void					camera_per_calc_piv(t_cam *cam);
 
 void					dots_fill_cb(const t_vec4 *p, const t_mtx4 *c,
 						t_vec3 *res);
 void					dots_fill_sb_iso(const t_vec3 *p, t_vec2 *res,
-						const float f);
+						const long double f);
 void					dots_fill_sb_per(const t_vec3 *point, t_vec2 *res,
 						const t_cam *cam);
 t_dots					*dots_init(const int x, const int y);
-void					dots_scale_y(t_dots *dots, const float s);
+void					dots_scale_y(t_dots *dots, const long double s);
 void					dots_color_update(t_dots *dots);
 
 void					image_clear(t_image *img);
@@ -216,15 +216,15 @@ t_mlx					*mlx_free_ll3d(t_mlx **mlx);
 
 void					oclude_x(t_vec2 *d1, t_vec2 *d2);
 void					oclude_y(t_vec2 *d1, t_vec2 *d2);
-void					oclude_z(t_vec2 *d1, t_vec2 *d2, float z1, float z2);
+void					oclude_z(t_vec2 *d1, t_vec2 *d2, long double z1, long double z2);
 void					oclusion_draw_line(const t_mlx *c, t_dots *d,
 						int d1, int d2);
 int						check_out(const t_dots *d, const int d1, const int d2);
 int						swap_if_d2_less(t_vec2 **d1, t_vec2 **d2,
-						const float d1v, const float d2v);
+						const long double d1v, const long double d2v);
 
-int						color_lerp(int c1, int c2, float step);
+int						color_lerp(int c1, int c2, long double step);
 int						color_get_default(const int z, const int cm);
-int						color_set_br(const int c, float br);
+int						color_set_br(const int c, long double br);
 
 #endif
