+++ ll3d/src/camera/camera_init.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   camera_init.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sbednar <sbednar@student.42.fr>            +#+  +:+       +#+        */
+/*   By: sbednar <sbednar@student.fr.42>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/04 05:39:55 by sbednar           #+#    #+#             */
-/*   Updated: 2019/02/14 20:36:13 by sbednar          ###   ########.fr       */
+/*   Updated: 2019/03/04 18:01:26 by sbednar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ll3d.h"
 
-static void	camera_init_1(float val[4][4])
+static void	camera_init_1(long double val[4][4])
 {
 	val[0][0] = 1;
 	val[0][1] = 0;
