+++ ll3d/src/camera/camera_rotate_y.c
@@ -6,13 +6,13 @@
 /*   By: sbednar <sbednar@student.fr.42>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/04 06:33:58 by sbednar           #+#    #+#             */
-/*   Updated: 2019/02/13 23:16:13 by sbednar          ###   ########.fr       */
+/*   Updated: 2019/03/04 18:01:26 by sbednar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ll3d.h"
 
-void	camera_rotate_y(t_cam *cam, const float a)
+void	camera_rotate_y(t_cam *cam, const long double a)
 {
 	t_mtx4	*tmp;
 	t_mtx4	*mr;
