+++ ll3d/src/color/color_lerp.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   color_lerp.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sbednar <sbednar@student.42.fr>            +#+  +:+       +#+        */
+/*   By: sbednar <sbednar@student.fr.42>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/11 19:42:55 by sbednar           #+#    #+#             */
-/*   Updated: 2019/02/14 20:37:33 by sbednar          ###   ########.fr       */
+/*   Updated: 2019/03/04 18:01:26 by sbednar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ll3d.h"
 
-int	color_lerp(const int c1, const int c2, const float step)
+int	color_lerp(const int c1, const int c2, const long double step)
 {
 	int	r;
 	int	g;
