+++ ll3d/src/color/color_set_br.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   color_set_br.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sbednar <sbednar@student.42.fr>            +#+  +:+       +#+        */
+/*   By: sbednar <sbednar@student.fr.42>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/13 23:49:12 by sbednar           #+#    #+#             */
-/*   Updated: 2019/02/14 20:37:18 by sbednar          ###   ########.fr       */
+/*   Updated: 2019/03/04 18:01:26 by sbednar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ll3d.h"
 
-int	color_set_br(const int c, float br)
+int	color_set_br(const int c, long double br)
 {
 	int	r;
 	int	g;
