+++ ll3d/src/dots/dots_fill_sb_iso.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   dots_fill_sb_iso.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sbednar <sbednar@student.42.fr>            +#+  +:+       +#+        */
+/*   By: sbednar <sbednar@student.fr.42>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/04 07:42:56 by sbednar           #+#    #+#             */
-/*   Updated: 2019/02/11 16:01:07 by sbednar          ###   ########.fr       */
+/*   Updated: 2019/03/04 18:01:26 by sbednar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ll3d.h"
 
-void	dots_fill_sb_iso(const t_vec3 *p, t_vec2 *res, const float f)
+void	dots_fill_sb_iso(const t_vec3 *p, t_vec2 *res, const long double f)
 {
 	res->x = (int)(p->x * f);
 	res->y = (int)(p->y * f);
