+++ ll3d/src/dots/dots_scale.c
@@ -6,13 +6,13 @@
 /*   By: sbednar <sbednar@student.fr.42>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/06 04:14:04 by sbednar           #+#    #+#             */
-/*   Updated: 2019/02/13 23:20:41 by sbednar          ###   ########.fr       */
+/*   Updated: 2019/03/04 18:01:26 by sbednar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ll3d.h"
 
-void	dots_scale_y(t_dots *dots, const float s)
+void	dots_scale_y(t_dots *dots, const long double s)
 {
 	int	i;
 	int	c;
