+++ ll3d/src/image/image_draw_line2.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   image_draw_line2.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sbednar <sbednar@student.42.fr>            +#+  +:+       +#+        */
+/*   By: sbednar <sbednar@student.fr.42>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/13 23:57:11 by sbednar           #+#    #+#             */
-/*   Updated: 2019/02/14 20:39:51 by sbednar          ###   ########.fr       */
+/*   Updated: 2019/03/04 18:01:26 by sbednar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,10 +24,10 @@ int	image_draw_line2(const t_mlx *mlx, int *t,
 		t[10] += t[9];
 		dt[4] = dt[2] >> dt[0];
 		image_set_pixel(mlx->image, t[10], t[12], color_set_br(color_lerp(
-			t[14], t[15], (float)ft_fabs(t[10] - t[16]) / ft_fabs(t[11] -
+			t[14], t[15], (long double)ft_fabs(t[10] - t[16]) / ft_fabs(t[11] -
 			t[16])), (255.0f - dt[4]) / 255.0f));
 		image_set_pixel(mlx->image, t[10], t[12] + 1, color_set_br(color_lerp(
-			t[14], t[15], (float)ft_fabs(t[10] - t[16]) / ft_fabs(t[11] -
+			t[14], t[15], (long double)ft_fabs(t[10] - t[16]) / ft_fabs(t[11] -
 			t[16])), (255.0f - (dt[4] ^ dt[5])) / 255.0f));
 	}
 	image_set_pixel(mlx->image, t[11], t[13], t[15]);
