+++ ll3d/src/math/ft_fabs.c
@@ -6,13 +6,13 @@
 /*   By: sbednar <sbednar@student.fr.42>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/04 03:15:14 by sbednar           #+#    #+#             */
-/*   Updated: 2019/02/13 23:21:58 by sbednar          ###   ########.fr       */
+/*   Updated: 2019/03/04 18:01:26 by sbednar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ll3d.h"
 
-float	ft_fabs(const float x)
+long double	ft_fabs(const long double x)
 {
 	return (x >= 0 ? x : -x);
 }
