+++ ll3d/src/math/ft_min_max.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_min_max.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: edraugr- <edraugr-@student.42.fr>          +#+  +:+       +#+        */
+/*   By: sbednar <sbednar@student.fr.42>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/11 22:24:33 by edraugr-          #+#    #+#             */
-/*   Updated: 2019/02/11 22:31:08 by edraugr-         ###   ########.fr       */
+/*   Updated: 2019/03/04 18:01:26 by sbednar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ll3d.h"
 
-float	ft_fmin(const float a, const float b)
+long double	ft_fmin(const long double a, const long double b)
 {
 	return (a > b ? b : a);
 }
@@ -22,7 +22,7 @@ int		ft_min(const int a, const int b)
 	return (a > b ? b : a);
 }
 
-float	ft_fmax(const float a, const float b)
+long double	ft_fmax(const long double a, const long double b)
 {
 	return (a < b ? b : a);
 }
