+++ ll3d/src/mtx4/mtx4_init.c
@@ -6,13 +6,13 @@
 /*   By: sbednar <sbednar@student.fr.42>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/17 22:58:25 by edraugr-          #+#    #+#             */
-/*   Updated: 2019/02/13 23:32:57 by sbednar          ###   ########.fr       */
+/*   Updated: 2019/03/04 18:01:26 by sbednar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ll3d.h"
 
-t_mtx4	*mtx4_init(const float val)
+t_mtx4	*mtx4_init(const long double val)
 {
 	t_mtx4	*res;
 	int		i;
